Memory_stage: first attempt
This is a first attempt of the memory stage of the pipeline.
The latency of this module is 3 clock cycles and stalls upstream for at
least 1 clk in case of store:
- loads: 1 cycle to set @, 1 cycle to get data, 1 cycle to register
outputs (and sign-extend/mask)
- store: 1 cycle to set @, data, we, 1 stall cycle to get wr ACK, 1 cycle to
register outputs.

Known issue: if the stage is killed while a store is processed, an
exception is raised, and stage is killed. This means that we don't
really know if the store happened, and this is not good.

TODO/optimization:
- set @/data directly at EX stage output
- concerning store, it can be done in WB stage, so the exception
handling is easier in case of kill. Moreover, as something is committed
in memory, this should be done in one single place (WB).
- For stores, we can reverse the logic and do a store per clk in case of
cache hit. Instead of ACK the store, a flag can be raised if the cache
miss or is busy. The data to store is registered somewhere in the cache,
so the algo for store could be: 1- stall if Busy (ie last store takes a
long time) 2- set @, data, we for 1 cycle. This technics partially hides
a cache miss.

Signed-off-by: Simon Desfarges <simon.desfarges@free.fr>