Caches: test: switch to auto-generated memory contents
Instead of reading from a file, generate an initial memory state, which
each data equals address + 0x0100. This is what is assumed by the
testbenches for now.

As part of the process, remove the local memory simulator, and switch to
the rjarzmik package generic one, which is reused across Caches and MIPS
Processor testers.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>